test(Input): Mask snapshot tests
